--- src/CMakeLists.txt
@@ -1,1 +1,0 @@
-/opt/ros/melodic/share/catkin/cmake/toplevel.cmake
