--- src/turtlesim_editor/action/turtlesim.action
@@ -1,9 +1,0 @@
-#goal definition
-float32 x
-float32 y
----
-#result definition
-duration timeTaken
----
-#feedback
-float32 distance
