--- src/turtlesim_editor/include/turtle/turtle.h
@@ -1,10 +1,0 @@
-#include <turtlesim/Pose.h>
-
-class turtle{
-  public:
-  float x;
-  float y;
-  float theta;
-  void callback(const turtlesim::Pose::ConstPtr& msg);
-};
-
