--- src/turtlesim_editor/launch/turtlesim_editor.launch
@@ -1,15 +1,0 @@
-<launch>
-
-  <group ns="turtlesim">
-    <node pkg="turtlesim" name="sim" type="turtlesim_node">
-        <param name = "background_r" value="128"  type="int"/>
-        <param name = "background_g" value="0"  type="int"/>
-        <param name = "background_b" value="128"  type="int"/>
-    </node>
-    <node pkg="turtlesim_editor" name ="turtlesim_clear" type = "turtlesim_clear_node"/>
-    <node pkg="turtlesim_editor" name ="turtlesim_spawn" type = "turtlesim_spawn_node"/>
-    <node pkg="turtlesim_editor" name ="turtlesim_moveback" type = "turtlesim_moveback_node"/>
-    <node pkg="turtlesim_editor" name ="turtlesim_distance" type = "turtlesim_distance_node"/>
-    <node pkg="turtlesim_editor" name ="turtlesim_move" type = "turtlesim_move_node"/>
-  </group>
-</launch>
