--- src/turtlesim_editor/msg/distance.msg
@@ -1,3 +1,0 @@
-float32 x
-float32 y
-float32 total
