--- src/turtlesim_editor/srv/moveback.srv
@@ -1,2 +1,0 @@
----
-bool moveSuccess
